working on building connection

+++ mongodb/mongodb/src/main/java/com/example/mongodb/User/UserService.java
@@ -3,14 +3,15 @@ package com.example.mongodb.User;
 import java.time.LocalDate;
 import java.util.List;
 
+import org.springframework.data.mongodb.repository.cdi.MongoRepositoryBean;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UserService {
+public class UserService{
     public List<User> getUsers(){
         return List.of(
             new User(
-                    (long) 1,
+                    1,
                     "Fonix",
                     "email@email.com",
                     LocalDate.now()
